Further simplify leader election by removing locks
CuratorElectionService is already thread safe; further, it uses locks.
By using also using locks in our election service, we were burdened
with the complexity of escaping them in order to avoid
deadlocks (which happened).

Further, we were calling in to external objects that we do not
control while we had said lock. While in theory this should have been
fine, it is an additional source of hidden complexity.

We can easily use AtomicBooleans to coordinate as needed. Since state
transitions only one direction, this becomes easy and manageable.